% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.regexport.coeftest.R
\name{as.regexport.coeftest}
\alias{as.regexport.coeftest}
\title{Coerce coeftest to class regexport}
\usage{
\method{as.regexport}{coeftest}(coeftest, model, sumstats = NULL)
}
\arguments{
\item{coeftest}{An object of class coeftest}

\item{model}{The lm or glm model from which the coeftest was calculated}

\item{sumstats}{Optional named list of additional summary statistics}
}
\value{
An object of class regexport
}
\description{
Apply revised standard errors from a coeftest (from package lmtest) to
a model and coerce into class regexport
}
