% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexcel.R
\name{regexcel}
\alias{regexcel}
\title{Output a regression to Excel}
\usage{
regexcel(
  reglist,
  file = "output.xlsx",
  sheet = "Regression",
  order = NULL,
  altnames = NULL,
  regnames = NULL,
  suppress = NULL,
  digits = 3,
  sumstats = c("R-squared", "Adj. R-squared", "Observations"),
  notes = "*** p<0.01, ** p<0.05, * p<0.1",
  siglevels = c(0.1, 0.05, 0.01),
  coefstyles = NULL,
  sumstyles = NULL,
  sestyle = NULL,
  startrow = 1
)
}
\arguments{
\item{reglist}{Regression output or a list of regression outputs that can be
coerced into class reglist}

\item{file}{An openxlsx workbook or the name of a new Excel file to create 
where output should be stored}

\item{sheet}{The name of the worksheet where results should be saved}

\item{order}{The order of variables in the coefficient list; any variables 
not in this list will be ordered automatically according to their 
order in the input regression models}

\item{altnames}{A named character vector giving the alternative names for
input variables; the vector results will replace the names}

\item{regnames}{An optional vector with names for each regression to place
on the row above the variable names (if NULL, this line will be 
omitted).}

\item{suppress}{A regular expression or vector of regressors to be 
suppressed from the output (useful for fixed effects)}

\item{siglevels}{What are the significance levels for which special 
formatting should be used?}

\item{sigstyles}{What styles should be used for each of the significance
levels? Defaults to typical stars used in regression tables}

\item{note}{Any notes to include at the bottom. Using "\%stars\%" will 
include the levels used for significance testing if the default
stars are used}
}
\description{
Output a regression to Excel
}
\details{
The \code{regexcel} function is used to output regressions to xlsx
format for presentation or further analysis in Microsoft Excel or other 
software that can read xlsx files. Because many users will wish to continue
their analysis in Excel, regexcel produces output that preserves numerical
formatting, but still produces the typical stars
}
\examples{
data <- data.frame(x = c(2, 4, 5, 6, 8, 10), y = c(1, 1, 2, 2, 3, 3))
regression <- lm(y ~ x, data = data)
regexcel(regression)
}
