% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexcel - old.R, R/regexcel.R
\name{regexcel}
\alias{regexcel}
\title{Save gravity model output as an xlsx file.}
\usage{
regexcel(
  reglist,
  file = "output.xlsx",
  sheet = "Regression",
  order = NULL,
  altnames = NULL,
  regnames = NULL,
  suppress = NULL,
  siglevels = c(0.1, 0.05, 0.01),
  stylecoef = NULL,
  stylese = NULL,
  stylesum = NULL,
  note = "\%stars\%"
)

regexcel(
  reglist,
  file = "output.xlsx",
  sheet = "Regression",
  order = NULL,
  altnames = NULL,
  regnames = NULL,
  suppress = NULL,
  siglevels = c(0.1, 0.05, 0.01),
  stylecoef = NULL,
  stylese = NULL,
  stylesum = NULL,
  note = "\%stars\%"
)
}
\arguments{
\item{reglist}{Regression output or a list of regression outputs that can be
coerced into class reglist}

\item{file}{An openxlsx workbook or the name of a new Excel file to create 
where output should be stored}

\item{sheet}{The name of the worksheet where results should be saved}

\item{order}{The order of variables in the coefficient list; any variables 
not in this list will be ordered automatically according to their 
order in the input regression models}

\item{altnames}{A named character vector giving the alternative names for
input variables; the vector results will replace the names}

\item{regnames}{An optional vector with names for each regression to place
on the row above the variable names (if NULL, this line will be 
omitted).}

\item{suppress}{A regular expression or vector of regressors to be 
suppressed from the output (useful for fixed effects)}

\item{siglevels}{What are the significance levels for which special 
formatting should be used?}

\item{note}{Any notes to include at the bottom. Using "\%stars\%" will 
include the levels used for significance testing if the default
stars are used}

\item{reg}{The results from lm or glm}

\item{wbname}{The name of the workbook to save the results to}

\item{reg.names}{A list of names for the regressions}

\item{coeftest}{The results of coeftest, if applicable}

\item{stars}{What levels should be used for the creation of stars?}

\item{alt.names}{A named character vector giving the alternative names for input variables; the vector results will replace the names}

\item{open}{Should the file be opened (using the openXL function) upon completion?}

\item{sigstyles}{What styles should be used for each of the significance
levels? Defaults to typical stars used in regression tables}
}
\description{
Save gravity model output as an xlsx file.

Output a regression to Excel
}
\details{
This is a utility function to output regression results as an Excel file. It is not used in the website and may eventually be moved to
another library.

The \code{regexcel} function is used to output regressions to xlsx
format for presentation or further analysis in Microsoft Excel or other 
software that can read xlsx files. Because many users will wish to continue
their analysis in Excel, regexcel produces output that preserves numerical
formatting, but still produces the typical stars
}
\examples{
data <- data.frame(x = c(2, 4, 5, 6, 8, 10), y = c(1, 1, 2, 2, 3, 3))
regression <- lm(y ~ x, data = data)
reg.to.xlsx(regression)
data <- data.frame(x = c(2, 4, 5, 6, 8, 10), y = c(1, 1, 2, 2, 3, 3))
regression <- lm(y ~ x, data = data)
reg.to.xlsx(regression)
}
